improved worker to use O(1) memory instead of O(n)
Previous approach used a queue of prefixes that will grow super large
the longer the worker runs.

Instead, for each worker, we define a starting number and a hop value.
Each worker attempts the start value password, then hops forward by the
hop value. When a digit overflows, we just carry the value and continue.